Proyectos/Asistente SAP/requirements.txt
# FastAPI y servidor
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Base de datos
sqlalchemy==2.0.23
alembic==1.12.1
psycopg[binary]==3.1.12
asyncpg==0.29.0

# Vector DB
qdrant-client==1.6.9

# ML y embeddings
openai==1.3.6
tiktoken==0.5.2
sentence-transformers==2.2.2

# Procesamiento de archivos
pdfminer.six==20221105
python-docx==1.1.0
beautifulsoup4==4.12.2
markdown==3.5.1
unstructured==0.11.6

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Tareas programadas
apscheduler==3.10.4

# Utilidades
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
structlog==23.2.0
rich==13.7.0

# Rate limiting
slowapi==0.1.9

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Linting y formato
ruff==0.1.6
black==23.11.0
mypy==1.7.1

# Monitoring
prometheus-client==0.19.0
